Changed from ushort to int

+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/EndUser/Dialogs/Queries/Get/DialogDto.cs
@@ -136,12 +136,12 @@ public sealed class DialogDto
     /// <summary>
     /// Indicates how many transmissions are not of type Submission or Correction
     /// </summary>
-    public ushort IncomingTransmissions { get; set; }
+    public int IncomingTransmissions { get; set; }
 
     /// <summary>
     /// Indicates how many transmissions are of type Submission or Correction
     /// </summary>
-    public ushort OutgoingTransmissions { get; set; }
+    public int OutgoingTransmissions { get; set; }
 
     /// <summary>
     /// The attachments associated with the dialog (on an aggregate level).
