+++ src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Get/DialogDto.cs
@@ -146,12 +146,12 @@ public sealed class DialogDto
     /// <summary>
     /// Indicates how many transmissions are not of type Submission or Correction
     /// </summary>
-    public ushort IncomingTransmissions { get; set; }
+    public int IncomingTransmissions { get; set; }
 
     /// <summary>
     /// Indicates how many transmissions are of type Submission or Correction
     /// </summary>
-    public ushort OutgoingTransmissions { get; set; }
+    public int OutgoingTransmissions { get; set; }
 
     /// <summary>
     /// The list of words (tags) that will be used in dialog search queries. Not visible in end-user DTO.
